pulled out all calcs, exposed renaming, streamlined download to csv process
