+++ expworkup/handlers/calc_conc.py
@@ -1,0 +1,174 @@
+import logging
+
+modlog = logging.getLogger(__name__)
+
+def reag_info(reagentdf,chemdf):
+    '''
+        ### Eventually this section of code can be replaced by class objects which describe the reagent for each experiment in line
+        ### Currenlty this code is assembling a dataframe which describes the relatinship between various reagents in order
+        ### to perform subsequent calcualtions.  The inchi keys for example are hard coded and should be variable
+        #TODO use types from teh chemdf to handle each of these sections with exceptions for parsing
+    '''
+    reagentlist=[]
+    for item in list(reagentdf):
+        name='null'
+        mm='null'
+        density='null'
+        density_new='null'
+        units='null'
+        if ('_raw_reagent_' in item) and ('InChIKey' in item):
+            for InChIKey in reagentdf[item]:
+                m_type='null'
+                parse_name=item[:-21]
+                parse_chemical_name=item[:-9]
+                # list of acids
+                if InChIKey == "BDAGIHXWWSANSR-UHFFFAOYSA-N":
+                    mm=(float(chemdf.loc[InChIKey,"Molecular Weight (g/mol)"]))# / float(chemdf.loc["FAH","Density            (g/mL)"])
+                    name=((chemdf.loc[InChIKey,"Chemical Name"]))# / float(chemdf.loc["FAH","Density            (g/mL)"])
+                    m_type='pureliquid' #Assigns an object type for later analysis and calculation of concentrations
+                    try:
+                        density=(float(chemdf.loc[InChIKey,"Density            (g/mL)"]))
+                    except:
+                        modlog.error("Abort run and check %s density details in google sheets" %name)
+                        break
+                # list of all solvents
+                elif InChIKey == 'YEJRWHAVMIAJKC-UHFFFAOYSA-N' \
+                        or InChIKey == 'ZMXDDKWLCZADIW-UHFFFAOYSA-N' \
+                        or InChIKey == 'IAZDPXIOMUYVGZ-UHFFFAOYSA-N' \
+                        or InChIKey == 'YMWUJEATGCHHMB-UHFFFAOYSA-N' \
+                        or InChIKey == 'MVPPADPHJFYWMZ-UHFFFAOYSA-N' \
+                        or InChIKey == 'UserDefinedSolvent':
+                    try: 
+                        mm=(float(chemdf.loc[InChIKey,"Molecular Weight (g/mol)"]))
+                    except Exception:
+                        mm=chemdf.loc[InChIKey,"Molecular Weight (g/mol)"]
+                    name=((chemdf.loc[InChIKey,"Chemical Name"]))# / float(chemdf.loc["FAH","Density            (g/mL)"])
+                    m_type='solvent'
+                    try:
+                        density=(float(chemdf.loc[InChIKey,"Density            (g/mL)"]))
+                    except:
+                        modlog.error("Abort run and check %s density details in google sheets" %name)
+                        break
+                # list of inorganics
+                elif InChIKey == 'RQQRAHKHDFPBMC-UHFFFAOYSA-L' \
+                        or InChIKey == 'ZASWJUOMEGBQCQ-UHFFFAOYSA-L':
+                    mm=(float(chemdf.loc[InChIKey,"Molecular Weight (g/mol)"]))
+                    name=((chemdf.loc[InChIKey,"Chemical Name"]))# / float(chemdf.loc["FAH","Density            (g/mL)"])
+                    m_type='inorg'
+                    try:
+                        density=(float(chemdf.loc[InChIKey,"Density            (g/mL)"]))
+                    except:
+                        modlog.error("Abort run and check %s density details in google sheets" %name)
+                        break
+                #Null
+                elif InChIKey == 'null':
+                    pass
+                #Organic
+                else:
+                    mm=(float(chemdf.loc[InChIKey,"Molecular Weight (g/mol)"]))
+                    name=((chemdf.loc[InChIKey,"Chemical Name"]))
+                    m_type='org'
+                    try:
+                        density = (float(chemdf.loc[InChIKey,"Density            (g/mL)"]))
+                    except:
+                        modlog.error("Abort run and check %s density details in google sheets" %name)
+                        break
+                for (item) in list(reagentdf):
+                    if (parse_chemical_name in item) and ("actual_amount" in item) and not ("units" in item):
+                        amount=(reagentdf.loc[0,item])
+                    if (parse_chemical_name in item) and ("actual_amount" in item) and ("units" in item):
+                        units=(reagentdf.loc[0,item])
+            reagentlist.append((name, mm, density, parse_name, InChIKey, m_type, amount, units))
+    reagentlist_df=(pd.DataFrame(reagentlist, columns=['name', 'molecular mass', 'density', 'parsed name', "InChIKey", "m_type", "amount", "units"]))
+    return(reagentlist_df)
+
+#Organizes all of the gathered reagent data by reagent and proceeds to send off for calculating molality                        
+def ReagentConc(one_reagent_df, reagent):
+    df=one_reagent_df
+    index=0
+    conc={}
+    reagent=reagent[5:]
+    #Calculateds the concentration of the REAGENTS being used in each experiment (signified largely by a single runID_vial combination)
+    for m_type in df['m_type']:
+        if m_type == 'null':
+            pass
+        if m_type == 'solvent':
+            try:
+                solv_mm=float(df.loc[(index,'molecularmass')])
+                solv_d=float(df.loc[(index,'density')])
+            except Exception:
+                solv_mm=(df.loc[(index,'molecularmass')])
+                solv_d=(df.loc[(index,'density')])
+            solv_A=float(df.loc[(index,'amount')])
+        index+=1
+    index=0
+    #calculates the values for the reagent concentrations in terms of the total solvent used in the preparation (for inorganic and organic components)
+    for m_type in df['m_type']:
+        if (m_type == 'inorg') or (m_type == 'org'):
+            mm=float(df.loc[(index,'molecularmass')])
+            A=float(df.loc[(index,'amount')])
+            calculated_concentration=(A / mm ) / ( solv_A / 1000) ## Returns the value in mmolarity
+#            name_space=('_rxn_' + reagent + "_chemical_" + str(index) + "_conc_" + (df.loc[(index,'InChiKey')]))
+            name_space=('_raw_' + reagent + "_conc_" + (df.loc[(index,'InChiKey')]))
+            conc[name_space]=calculated_concentration
+        #Calculates the pure "density" of the compound and uses that in place of the solvated concentration using the amounts from the experiment
+        if (m_type == 'pureliquid'):
+            mm=float(df.loc[(index,'molecularmass')])
+            A=float(df.loc[(index,'amount')])
+            d=float(df.loc[(index,'density')])
+            calculated_concentration=(d/mm*1000)
+            name_space=('_raw_' + reagent + "_conc_" + (df.loc[(index,'InChiKey')]))
+            conc[name_space]=calculated_concentration
+        index+=1
+    if conc == {}:
+        return('null')
+    else:
+        return([conc])
+
+def calcConc(reagentdf, reagent_spc):
+    """ calculates concentration of reagents
+
+    TODO: Move conc_cells calculation into ReagentObject Class
+    TODO: move organization of reagent specification sheet from chemical dataframe into separate function
+    TODO: single step organization of dataframes for concentration calculation
+    TODO: LONG TERM -- reagent class constructed at initial JSON parsing
+
+    :param reagentdf: dataframe of reagent information for a grouping of experiments
+    :param reagent_spc: dataframe generated from chemical inventory indexed by reagent number for a group of experiments
+    :return: dataframe of concentrations with columns sorted by inchikey and reagent
+                (ex. '_raw_reagent_1_conc_UPHCENSIMPJEIS-UHFFFAOYSA-N')
+    """
+    reagent_spc.set_index(['parsed name'], inplace=True)
+    reagent_list=[]
+    conc_df=pd.DataFrame()
+    for reag_chem, row in reagent_spc.iterrows():
+        if reag_chem not in reagent_list:
+            reagent_list.append(reag_chem)
+        else:
+            pass
+    count = 0
+    for reagent in reagent_list:
+        count += 1
+        for reag_chem, row in reagent_spc.iterrows():
+            if reagent in reag_chem:
+                chemical_list = []
+                chemical_list.append(reagent_spc.at[reag_chem, 'name'])
+                chemical_list.append(reagent_spc.at[reag_chem, 'InChIKey'])
+                chemical_list.append(reagent_spc.at[reag_chem, 'density'])
+                chemical_list.append(reagent_spc.at[reag_chem, 'm_type'])
+                chemical_list.append(reagent_spc.at[reag_chem, 'molecular mass'])
+                chemical_list.append(reagent_spc.at[reag_chem, 'amount'])
+                chemical_list.append(reagent_spc.at[reag_chem, 'units'])
+        onereagent_df = (pd.DataFrame(chemical_list).transpose())
+        onereagent_df.columns = ['name','InChiKey', 'density', 'm_type',
+                                    'molecularmass','amount','unit']
+        conc_cells = (ReagentConc(onereagent_df, reagent))
+        current_reagent = ReagentObject(onereagent_df, reagent)
+        v1conc_cells = current_reagent.conc_v1
+        if conc_cells == 'null':
+            pass
+        else:
+            conc_cells_df = pd.DataFrame(conc_cells)
+            v1conc_cells_df = pd.DataFrame(v1conc_cells)
+            conc_df = pd.concat([conc_df, conc_cells_df, v1conc_cells_df], axis=1)
+    return conc_df
